+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - hessenberg tranformer with givens rotations
 - householder hessenberg transformer
 - qr without shift method
+- exercise 5.4
 
 ## 0.4.4 - 2017-05-28
 ### Added